requirements.txt
langchain 
flask
pypdf
python-dotenv
sentence-transformers
pinecone[grpc]
langchain-pinecone
langchain_community
langchain-groq
langchain_experimental
groq
flask_session
-e .